+++ src/zope/testrunner/runner.py
@@ -1027,7 +1027,9 @@ class TestResult(unittest.TestResult):
 
     def stopTest(self, test):
         self.testTearDown()
-        self.options.output.stop_test(test)
+        gccount = gc.collect() if self.options.gc_after_test and not is_jython \
+                  else 0
+        self.options.output.stop_test(test, gccount)
 
         if is_jython:
             pass