+++ CHANGES.rst
@@ -5,7 +5,13 @@
 5.4.1 (unreleased)
 ==================
 
-- Nothing changed yet.
+- New option ``--gc-after-test``. It calls for a garbage collection
+  after each test and can be used to track down ``ResourceWarning``s
+  and cyclic garbage.
+  With ``rv = gc.collect()``, ``!`` is output on verbosity level 1 when
+  ``rv`` is non zero (i.e. when cyclic structures have been released)
+  and ``[``*rv*``]`` on higher verbosity levels
+  (`#131 <https://github.com/zopefoundation/zope.testrunner/issues/131`_).
 
 
 5.4.0 (2021-11-19)